Fix `cp -r` on macOS/FreeBSD (?) systems
cp -r behaves differently on macOS/FreeBSD(?) when the path ends with a trailing slash, see man page:
If source_file designates a directory, cp copies the directory and the entire subtree connected at that point. If the source_file ends in a /, the contents of the directory are copied rather than the directory itself.

On linux systems there is no difference.